+++ ts-node/src/interfaces/index.ts
@@ -1,0 +1,2 @@
+export { Hero } from '../classes/Hero';
+export { Villain } from './villain';
